🐞 fix: 修复购物车服务中的用户ID类型，确保一致性并优化数据处理

+++ app/frontend/biz/service/delete_cart.go
@@ -19,7 +19,7 @@ func NewDeleteCartService(c context.Context) *DeleteCartService {
 }
 
 func (s *DeleteCartService) Run(req *cart_page.DeleteCartReq) (map[string]interface{}, error) {
-	id, ok := s.ctx.Value("user_id").(uint32)
+	id, ok := s.ctx.Value("user_id").(int64)
 	if !ok {
 		return nil, errors.New("no user id in context")
 	}
@@ -40,7 +40,7 @@ func (s *DeleteCartService) Run(req *cart_page.DeleteCartReq) (map[string]interf
 	}
 
 	resp, err := cartClient.DeleteCart(s.ctx, &pbcart.DeleteCartReq{
-		UserId: id,
+		UserId: uint32(id),
 		Items:  items,
 	})
 	if err != nil {
